requirements.txt
amqp==5.0.9
asgiref==3.5.0
attrs==21.4.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
dateparser==1.1.0
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.4
dj-database-url==0.5.0
dj-rest-auth==2.2.3
Django==4.0.2
django-allauth==0.47.0
django-baton==2.3.0
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-filter==21.1
django-formtools==2.3
django-google-analytics-app==5.0.5
django-js-asset==2.0.0
django-polymorphic==3.1.0
django-tinymce==3.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
filelock==3.6.0
greenlet==1.1.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
jet-bridge-base==1.1.8
Jinja2==3.1.1
jsl-django-sitemap==1.2.0
jsonschema==4.4.0
kavenegar==1.1.2
kombu==5.2.3
MarkupSafe==2.1.1
oauthlib==3.2.0
openapi-codec==1.3.2
packaging==21.3
Pillow==9.0.1
pipenv==2022.3.28
platformdirs==2.5.1
polib==1.1.1
prompt-toolkit==2.0.9
psycopg2==2.9.3
pycparser==2.21
pycryptodome==3.14.1
PyJWT==2.3.0
pyotp==2.6.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.19.2
python-gettext==4.0
python3-openid==3.2.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
regex==2022.3.2
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
social-auth-core==4.2.0
soupsieve==2.3.1
SQLAlchemy==1.4.32
sqlparse==0.4.2
structlog==21.5.0
swagger-spec-validator==2.7.4
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
virtualenv==20.14.0
virtualenv-clone==0.5.7
waitress==2.1.1
wcwidth==0.2.5
whitenoise==6.0.0
wrapt==1.13.3
YubiOTP==1.0.0
